requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
Faker==22.6.0
idna==3.10
pillow==11.1.0
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
urllib3==2.3.0
nudenet==2.0.9
